refactor: 将用户信息操作抽离 Context 类
